WIP: Improve Sidebar UX
