+++ src/frontend/components/UI/Sidebar/components/CurrentDownload/index.tsx
@@ -1,5 +1,4 @@
 import React, { useContext, useEffect, useState } from 'react'
-import { Link } from 'react-router-dom'
 import LinearProgress from '@mui/material/LinearProgress'
 import Typography from '@mui/material/Typography'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
@@ -18,6 +17,9 @@ type Props = {
   runner: Runner
 }
 
+const EOS_APP_NAME = '98bc04bc842e4906993fd6d6644ffb8d';
+const EOS_APP_RUNNER = 'legendary';
+
 export default React.memo(function CurrentDownload({ appName, runner }: Props) {
   const [progress] = hasProgress(appName)
   const [gameTitle, setGameTitle] = useState('')
@@ -29,8 +31,8 @@ export default React.memo(function CurrentDownload({ appName, runner }: Props) {
       // Hack for EOS Overlay. Not sure if this can be done better
       let title
       if (
-        appName === '98bc04bc842e4906993fd6d6644ffb8d' &&
-        runner === 'legendary'
+        appName === EOS_APP_NAME &&
+        runner === EOS_APP_RUNNER
       ) {
         title = 'EOS Overlay'
       } else {
@@ -52,8 +54,7 @@ export default React.memo(function CurrentDownload({ appName, runner }: Props) {
   }
 
   return (
-    <>
-      <Link to={`/download-manager`} className="currentDownload">
+    <div className="currentDownload">
         <span className="statusIcon" title={`${getStatus()} - ${gameTitle}`}>
           <Badge
             badgeContent={`${Math.round(progress.percent ?? 0)}%`}
@@ -82,7 +83,6 @@ export default React.memo(function CurrentDownload({ appName, runner }: Props) {
             </Box>
           </Box>
         </div>
-      </Link>
-    </>
+    </div>
   )
 })
