+++ src/frontend/components/UI/Sidebar/components/QuitButton/index.tsx
@@ -27,7 +27,7 @@ const QuitButton: React.FC = () => {
   }
 
   return (
-    <button className="Sidebar__item" onClick={() => handleQuitButton()}>
+    <button className="Sidebar__item Sidebar__quit" onClick={() => handleQuitButton()}>
       <div className="Sidebar__itemIcon">
         <FontAwesomeIcon
           icon={faPowerOff}
