--- src/frontend/components/UI/Sidebar/components/SidebarLinks/index.css
@@ -1,35 +1,0 @@
-.SidebarLinks {
-  grid-area: links;
-  font-size: var(--text-lg);
-}
-
-.Sidebar__item.SidebarLinks__subItem,
-.Sidebar__item.SidebarLinks__subItem:hover {
-  padding-block: var(--space-3xs);
-  padding-inline: var(--space-xl) var(--sidebar-horizontal-padding);
-  font-size: var(--text-sm);
-  line-height: 20px;
-  white-space: break-spaces;
-}
-
-.Sidebar__item:hover {
-  background-color: var(--navbar-active-background);
-  transition: 0.1s;
-  padding-inline-start: 12px;
-}
-
-.Sidebar__item.SidebarLinks__subItem::before {
-  content: '• ';
-}
-
-.SidebarLinks__subItem + .Sidebar__item:not(.SidebarLinks__subItem) {
-  margin-top: 4px;
-}
-
-@media screen and (max-width: 730px) {
-  .Sidebar__item.SidebarLinks__subItem,
-  .Sidebar__item.SidebarLinks__subItem:hover {
-    padding: var(--space-3xs) var(--sidebar-horizontal-padding) var(--space-3xs)
-      var(--space-lg);
-  }
-}
